backend/requirements.txt
# DJMashAI Backend — Hackathon MVP
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0

# Audio analysis
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.0

# AI (Gemini)
google-genai>=1.0.0

# YouTube audio extraction
yt-dlp>=2024.1.0

# Stem separation (optional, for stem-aware "Preview (stems)" transitions)
# Use demucs on Python 3.12+ (spleeter is not compatible with 3.12)
demucs>=4.0.0

# Vocal phrase boundaries for clean transitions (replace, not overlap)
openai-whisper>=20231117
